requirements.txt
Pillow>=10.0.0
pytesseract>=0.3.10
SQLAlchemy>=1.4.0